+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
+### Added
+
+* `dx watch` support for detailed job metrics (cpu, memory, network, disk io, etc every 60s) 
+
 ## [349.0] - beta
 
 ### Added